Refactored Node test environments to share more
