+++ .vscode/launch.json
@@ -77,8 +77,8 @@
                 "--no-warnings",
                 "${workspaceRoot}/node_modules/mocha/lib/cli/cli.js",
                 "--require",
-                "src/node/inject-globals.ts",
-                "src/index.ts",
+                "src/node/inject-dev-environment.ts",
+                "src/node/index.ts",
                 "--timeout",
                 "10000",
                 "--grep",
@@ -103,8 +103,8 @@
             ],
             "args": [
                 "--require",
-                "src/node/inject-globals.ts",
-                "src/index.ts",
+                "src/node/inject-dev-environment.ts",
+                "src/node/index.ts",
                 "--timeout",
                 "10000",
                 "--grep",
