+++ integration-tests/tests/.mocharc.json
@@ -8,7 +8,7 @@
   "reporter": "@realm/mocha-reporter",
   "require": [
     "source-map-support/register",
-    "src/node/inject-globals.ts"
+    "src/node/inject-dev-environment.ts"
   ],
-  "spec": "src/index.ts"
+  "spec": ["src/node/index.ts"]
 }
